Moved some state down to <WordBox/> component
<WordBox/> contains state to keep track of the word that is being typed by the user.
Originally, the state for <WordBox/> was stored in context.
This causes an unnecessary rerender for the entire app instead of just the box row that is currently being worked on.
That state is now local to <WordBox/> and a ref was added in GameContext to track all the submitted/used words by the user.

Another change is that the <Keyboard/> component is now a child of <WordBox/>, allowing it to share the method that is used to change the local state of <WordBox/>.
Since <Keyboard/> is now a child of <WordBox/>, it will still rerender while you are editing a word; but that was its original behavior anyways so I think this is still better for the most part.

With these changes, the following are the components that was excluded from unnecessary rerender:
 - <Modal/>
 - all 5 <WordBox/> components that are not focused on (only 1 is active at any given moment),